rename: type.ts -> types.ts
