+++ src/components/Covid/CovidStatus/types.ts
@@ -1,4 +1,4 @@
 export interface ICovidStatusItem {
-  data: { icon: JSX.Element; title: string };
+  data: { icon: JSX.Element | null; title: string };
   end: string;
 }
